Move controller to pkg/controller/catalog
