+++ pkg/controller/catalog/Dockerfile
