+++ pkg/controller/catalog/README.md
