+++ pkg/controller/catalog/service_controller_test.go
