+++ pkg/controller/catalog/storage/storage_util.go
