+++ pkg/controller/catalog/storage/tpr_storage.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"log"
 
-	"github.com/kubernetes-incubator/service-catalog/controller/util"
-	"github.com/kubernetes-incubator/service-catalog/controller/watch"
 	scmodel "github.com/kubernetes-incubator/service-catalog/model/service_controller"
+	"github.com/kubernetes-incubator/service-catalog/pkg/controller/catalog/util"
+	"github.com/kubernetes-incubator/service-catalog/pkg/controller/catalog/watch"
 
 	// Need this for gcp auth
 	_ "k8s.io/client-go/1.5/kubernetes"
