+++ pkg/controller/catalog/util/util.go
