+++ pkg/controller/catalog/util/util_test.go
