+++ pkg/controller/catalog/watch/dynamic_resource_client.go
