+++ pkg/controller/catalog/watch/typedata/service_binding_sample.yaml
