+++ pkg/controller/catalog/watch/watcher.go
