+++ pkg/controller/catalog/watch/watcher_test.go
